Splits Ability into sepparate files
