+++ Danki2/Assets/Scripts/Abilities/Ability.cs
@@ -1,30 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using UnityEngine;
-
-public enum AbilityReference
-{
-    Slash,
-    ShieldBash,
-    Whirlwind,
-    Fireball,
-}
-
-public class AbilityContext
-{
-    public AbilityContext(
-        Actor owner,
-        Vector3 targetPosition
-    )
-    {
-        Owner = owner;
-        TargetPosition = targetPosition;
-
-    }
-
-    public Actor Owner { get; }
-    public Vector3 TargetPosition { get; }
-}
 
 public class Ability
 {
