+++ Danki2/Assets/Scripts/Abilities/AbilityContext.cs
@@ -1,0 +1,17 @@
+﻿using UnityEngine;
+
+public class AbilityContext
+{
+    public AbilityContext(
+        Actor owner,
+        Vector3 targetPosition
+    )
+    {
+        Owner = owner;
+        TargetPosition = targetPosition;
+
+    }
+
+    public Actor Owner { get; }
+    public Vector3 TargetPosition { get; }
+}
