+++ Danki2/Assets/Scripts/Abilities/AbilityContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08872ef0a045e374bb9aabd56e250133
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
