+++ Danki2/Assets/Scripts/Abilities/AbilityReference.cs
@@ -1,0 +1,7 @@
+﻿public enum AbilityReference
+{
+    Slash,
+    ShieldBash,
+    Whirlwind,
+    Fireball,
+}
