+++ Danki2/Assets/Scripts/Abilities/AbilityReference.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a7ce49380115994b92ecc110b059a05
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
